MSPApp/etc: fix overflow bug causing us to start writing from the beginning of the SD card after only 720 images
This was caused a uint32_t address overflowing, because (720*(32+2304*1296*2)) = 4,299,840,000 > 0xFFFFFFFF. We fixed this by using SD card 512-byte block indexes everywhere, instead of using 32-bit addresses, since we give the SD card a 512-byte block index anyway, and never actually use the address itself.

+++ Code/Shared/MSP.h
@@ -15,11 +15,11 @@ namespace MSP {
     // ImgRingBuf: stats to track captured images
     struct [[gnu::packed]] ImgRingBuf {
         struct [[gnu::packed]] {
-            Img::Id idBegin     = 0;
-            Img::Id idEnd       = 0;
-            uint16_t widx       = 0;
-            uint16_t ridx       = 0;
-            bool full           = false;
+            Img::Id idBegin = 0;
+            Img::Id idEnd   = 0;
+            uint32_t widx   = 0;
+            uint32_t ridx   = 0;
+            bool full       = false;
         } buf;
         volatile bool valid = false;
